Fix akward Content-Type header.
It turns out that the environment add ';charset=ISO-8859-1' to the end
of the Content-Type header when HttpSocketResponse.getOutputWriter() is
called, creating the quite akward and nonsensical Content-Type of
'application/x-portable-graymap;charset=ISO-8859-1'. I now work around
this by using HttpSocketResponse.getOutputStream() and wrap the returnd
stream in a PrintWriter myself.